Vim: Disable reloading of vimrc
Vimrc automatic reload when saved is not working properly, it seems like
the colorscheme changes when the file is automatically reloaded.

Disabling this until I figure out what the problem is.

Also Improved the comment on wildoptions and replaced R (for row)
with L (for line) in the status line.