Vim: Have a separate local bundles file
When the local configuration is sourced, if it includes Bundle commands,
the file type detection needs to be turned on again but it doesn't seem
to work properly.

As a workaround, although I don't have a problem leaving it like this, I
am sourcing a specific local bundles.vim file at then end of the Bundle
definitions in vimrc. This way, file type detection is turned on only
once and the problem disappears.

I am still sourcing a local configuration file near the end of vimrc, so
I have moved those two to a directory under .vim called local. Bundle
commands in this second file are not guaranteed to work properly.