Screen: Use day/month date format in status line
I've started to use screen in a full screen terminal and, for the first
time in a long while, I've been looking at the time and date in the
status line. Turns out all this time the date has been set in month/day
format, which is not the natural way I read dates.