Screen: Add basic mouse wheel support to screen
Note that this does not work in copy mode, it simply scrolls up the
visible screen buffer. Changing windows within a screen sessions messes
this up, so scrolling up will show a mixture of the output in the
current window and the previous ones. For the time being, this is good
enough for me.