Bash: Load all functions first
Function dependencies are only important when the functions are
executed, not when they are created, so there is no need to come up with
a function order. I'm simply moving them to the beginning by prefixing
all files with functions with a 010. This will save me the headache of
figuring out which file to include first based on dependencies that
never existed.