Vim: Re-enable reloading of vimrc when modified
The autocommand to reload vimrc or any known configuration automatically
when those files are saved was disabled in db2fbfb7. This changeset
re-enables it, with some additions:

 * The edited file is reloaded, together with vimrc and gvimrc if
   running the gui.
 * Filetype detection is re-run after the config file has been reloaded.

This seems to work and colorschemes are still applied after reloading,
which was the problem that made me disable this feature in the first
place.