Bash: Add common ls typos as aliases for ls
Sometimes I type l and hit enter before typing the s, but I still type
it, and it is in the next command line when I type ls immediately after,
resulting in l, then sls. When this happens twice in a row, I have l,
then sl, then sls. These aliases are here to avoid the frustration that
ensues.