Vim: Add function to fill buffer with shell output
I use this with some tools that pipe some commend to a instance or
readonly vim (view) and add a command line map that calls this function
with the command itself. e.g.:

ls -l | view - +"map <F5> :call ReloadCommandOutput('ls -l')<CR>"

This will show the results of `ls -l` and will update them when F5 is
pressed.