Bash: Capture pipestatus in bash_prompt command
Provide a copy of the original pipestatus and return code of the command
that was just executed. This is so that the prompt theme functions can
react to the status of any commands in the pipe. These statuses, if not
captured, will be overwriten by the commands in the prompt handling
functions.