Bash: Add edot function to choose dotfiles to edit
Named after the use I intend to give it, but it can potentially be used
to edit any source files.

Uses *Pygments* for syntax highlighting the preview and vim to finally
edit. The editing happens all without ever exiting `fzf`, so that more
than one file can  be edited in the same session.