Bash: Acount for login shells when reloading a shell
I use reloadsh to refresh the config, but iTerm2 uses login shells by
default, and running reloadsh ends up with a partial configuration
applied (since eventually it does exec bash).

This change checks whether the current shell is a login shell and, if it
is, execs bash with the -l flag to start it as login shell as well.