Tmux: Disable aggressive-resize option
`aggresive-resize` is helpful but in macos, when using iTerm2 with the
Tmux integration, you get an error if the option is on:

-----------------------
Unsupported tmux option
---------------------------------------------------------------------------
The "aggressive-resize" option is turned on in tmux. It is not
compatible
with the iTerm2-tmux integration. Please disable it and try again.
---------------------------------------------------------------------------

I does not seem like it will be changed:
https://gitlab.com/gnachman/iterm2/-/issues/2585