Bash: Make edot more flexible
Edot: the tool to browse my dotfiles and open them for editing, now
supports multiple options for highlighted previews, starting with `bat`
if available and going through pygmentize, which was the only option
before, falling back to cat if none of the others are available.

This change also uses the editor set in the $EDITOR env var to open the
files.